Adds two CLI options: 1. --is-bare-expo-workflow Bare workflow expo apps want to update *everything*. This will update the app.json values and the /android and the /ios values.
2. --is-self-hosting-bundles
This one is a bit more esoteric, but it's meant for bare workflow apps using sentry-expo to upload their sourcemaps in a postExport hook, where it's essential to keep the release/dist versions align with the build/release versions. If it's passed in, it will look in the hooks inside the app.json file to update these values as needed.